requirements.txt
asgiref==3.2.10
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
fuzzywuzzy==0.18.0
jdcal==1.4.1
numpy==1.18.5
openpyxl==3.0.3
pandas==1.0.5
pdf2image==1.13.1
Pillow==7.1.2
psycopg2==2.8.5
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wfastcgi==3.0.0
